server: don't try to do nat traversal for a player talking to themselves
When there is only a single player left in a game, they form a ring with
themselves. Since they will receive a message back from the same port
they are talking to, we don't need to do NAT traversal. They actually
won't reply to a packet id 6. Maybe because they're the only player in
the game?